Authorize+View+small bugs

+++ Proiect2020/Controllers/CommentsController.cs
@@ -1,4 +1,5 @@
-﻿using Proiect2020.Models;
+﻿using Microsoft.AspNet.Identity;
+using Proiect2020.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -24,22 +25,32 @@ namespace Proiect2020.Controllers
         public ActionResult Delete(int id)
         {
             Comment comm = db.Comments.Find(id);
-            db.Comments.Remove(comm);
-            db.SaveChanges();
-            TempData["message"] = "Comentariul a fost sters";
-            return Redirect("/Tasks/Show/" + comm.TaskId);
+            if (comm.UserId == User.Identity.GetUserId() || User.IsInRole("Admin"))
+            {
+                db.Comments.Remove(comm);
+                db.SaveChanges();
+                TempData["message"] = "Comentariul a fost sters";
+                return Redirect("/Tasks/Show/" + comm.TaskId);
+            }
+            else
+            {
+                TempData["message"] = "Nu aveti dreptul sa stergeti comentariul";
+                return Redirect("/Tasks/Show/" + comm.TaskId);
+            }
         }
         [Authorize(Roles = "Membru,Organizator,Admin")]
         public ActionResult New(int id)
         {
-
+            Comment comm = new Comment();
             ViewBag.IdTask = id;
+            comm.UserId = User.Identity.GetUserId();
             return View();
         }
         [HttpPost]
         [Authorize(Roles = "Membru,Organizator,Admin")]
         public ActionResult New(Comment comm)
         {
+            comm.UserId = User.Identity.GetUserId();
             comm.Date = DateTime.Now;
             try
             {
